+++ lib/pdk/validate/metadata/task_metadata_lint.rb
@@ -1,6 +1,6 @@
 require 'pdk'
 require 'pdk/cli/exec'
-require 'pdk/validators/base_validator'
+require 'pdk/validate/base_validator'
 require 'pdk/util'
 require 'pathname'
 require 'json-schema'
