+++ lib/pdk/validate/puppet/puppet_lint.rb
@@ -1,7 +1,7 @@
 require 'pdk'
 require 'pdk/util'
 require 'pdk/cli/exec'
-require 'pdk/validators/base_validator'
+require 'pdk/validate/base_validator'
 
 module PDK
   module Validate
